Fix & extend variable parsing
The previous use of 'take_while' could silently match nothing without
causing issue. We extend it to take_while_m_n so that we definitely
match at least one lower case letter at the start of the variable name.
That broken lots of things so we introduce module & type constructor
checks on top of that to resolve the issues.