Break out more components & add more tests
We start to include tests for lists & tuples which seem to be working. I
don't fully understand the elm-ast 'binary' parser and I hope that we
don't need the operator table & precendence for what we're doing as
we're not interested in execution order for our initial task of writing
an elm linter.