Change approach to spaces_and_new_lines
So that we can try to accommodate single line comments in there.

We borrow the approach from the elm-syntax library which is to use
'many1' to progressively parse the possible whitespace whilst allowing
comments in there.

My first attempt to included something like many0(' ' | '\n') >> '\n'
but the parsers are greedy so that cannot work.

I've introduce some controls for equal, 'greater' or 'greater than or
equal' indents but I've not used anything other than GTE yet. I suspect
they'll be necessary but things are working at the moment.