Update newline & app parsing strategies
We change the application parsing strategy to require a 'greater than'
indent after a new line rather than 'gte' as otherwise we continue to
parse the next line that could be a new assignment in a let block.

We change the spaces_and_new_lines strategy to only care about the
indentation checks if there has been a new line. If it has only been
spaces then we can preserve the old new line and avoid validating the
indentation.