Extract function parsing
To separate functions & data types so that we can easily include them in
both the let-block parsing & top level functions. Previously we couldn't
handle function signatures in let-blocks. Now we distinguish between
functions which can have signatures and bindings which can't. I believe
this is roughly correct but I don't have anything offical to go on.