All enabled tests passing
With the 'combine' library. Eventually figured out how to get stuff
compiling. Not sure how long I'll remember it for or even what the
tricks were.

Though one of them was to stop using 'take_while' and use
'many(satisfy)' instead. The former produces a 'Range' type whilst the
latter allows you to collect the data into a value fairly easily which
is a relief.